requirements.txt
flask
Jinja2
joblib
gunicorn
kiwisolver
MarkupSafe
matplotlib
pandas
numpy
sklearn

Pillow
pipreqs
pyparsing
python-dateutil
pytz
requests
opencv-python-headless

six

threadpoolctl
urllib3
Werkzeug
yarg




keras
tensorflow-cpu

